pyproject.toml
[project]
name = "ahc-problem-digests"
version = "0.1.0"
description = "AHC の問題文を Gemini API で要約して保存するツール"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-genai>=1.73.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]

[project.scripts]
ahc-digest = "ahc_problem_digests.cli:main"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-mock>=3.15.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ahc_problem_digests"]
